Разделил файлы, остальное назметно
