+++ Source/cScreen.hpp
@@ -1,8 +1,9 @@
 #pragma once
+#include <SFML/Graphics.hpp>
 
 class cScreen
 {
 public:
     virtual int Run (sf::RenderWindow &App) = 0;
-    virtual ~cScreen();
+    virtual ~cScreen() {};
 };
