+++ Source/screens.hpp
@@ -1,3 +1,4 @@
+#include <iostream>
 const std::string RESOURCES_PATH_PREFIX =  "../Resources/Fonts/";
 
 #ifndef SCREENS_HPP_INCLUDED
